Make quantity a byte.

+++ src/main/java/com/assurity/coffee/CoffeeOrder.java
@@ -4,11 +4,11 @@ public class CoffeeOrder implements Comparable {
 
     private int customerType;
     private int productCode;
-    private int quantity;
+    private byte quantity;
     private CoffeeSize size;
     private int timeStamp;
 
-    public CoffeeOrder(int customerType, int productCode, int quantity, CoffeeSize size, int timeStamp) {
+    public CoffeeOrder(int customerType, int productCode, byte quantity, CoffeeSize size, int timeStamp) {
         this.customerType = customerType;
         this.productCode = productCode;
         this.quantity = quantity;
@@ -36,6 +36,7 @@ public class CoffeeOrder implements Comparable {
         CoffeeOrder other = (CoffeeOrder)o;
         int thisPriority = getCustomerPriority();
         int otherPriority = other.getCustomerPriority();
+        // If the customer type priorities are the same then the coffee order with the earlier timestamp takes priority.
         if (thisPriority == otherPriority) {
             return this.timeStamp - other.timeStamp;
         } else {
